refactor(utils): move standalone constants away
