+++ ggshield/cmd/iac/scan/prereceive.py
@@ -18,8 +18,7 @@ from ggshield.cmd.utils.common_options import all_option
 from ggshield.core.config import Config
 from ggshield.core.errors import handle_exception
 from ggshield.core.git_hooks.prereceive import get_breakglass_option, parse_stdin
-from ggshield.core.utils import EMPTY_TREE
-from ggshield.utils.git_shell import check_git_ref, is_valid_git_commit_ref
+from ggshield.utils.git_shell import EMPTY_TREE, check_git_ref, is_valid_git_commit_ref
 from ggshield.verticals.iac.collection.iac_diff_scan_collection import (
     IaCDiffScanCollection,
 )
