+++ ggshield/core/git_hooks/ci/commit_range.py
@@ -4,8 +4,7 @@ from typing import List, Tuple
 import click
 
 from ggshield.core.errors import UnexpectedError
-from ggshield.core.utils import EMPTY_SHA
-from ggshield.utils.git_shell import get_list_commit_SHA
+from ggshield.utils.git_shell import EMPTY_SHA, get_list_commit_SHA
 
 from .previous_commit import (
     github_pull_request_previous_commit_sha,
