+++ ggshield/core/git_hooks/ci/previous_commit.py
@@ -4,8 +4,7 @@ from typing import Optional
 import click
 
 from ggshield.core.errors import UnexpectedError
-from ggshield.core.utils import EMPTY_SHA
-from ggshield.utils.git_shell import get_last_commit_sha_of_branch, git
+from ggshield.utils.git_shell import EMPTY_SHA, get_last_commit_sha_of_branch, git
 
 from .supported_ci import SupportedCI
