+++ ggshield/core/git_hooks/prepush.py
@@ -4,8 +4,7 @@ import sys
 from typing import List, Optional, Tuple
 
 from ggshield.core.text_utils import display_warning
-from ggshield.core.utils import EMPTY_SHA
-from ggshield.utils.git_shell import git, is_valid_git_commit_ref
+from ggshield.utils.git_shell import EMPTY_SHA, git, is_valid_git_commit_ref
 
 
 logger = logging.getLogger(__name__)
