+++ tests/unit/cmd/scan/test_prepush.py
@@ -5,10 +5,13 @@ import pytest
 from click.testing import CliRunner
 
 from ggshield.__main__ import cli
+from ggshield.cmd.secret.scan.secret_scan_common_options import (
+    IGNORED_DEFAULT_WILDCARDS,
+)
 from ggshield.core.errors import ExitCode
 from ggshield.core.filter import init_exclusion_regexes
 from ggshield.core.scan import ScanContext, ScanMode
-from ggshield.core.utils import EMPTY_SHA, EMPTY_TREE, IGNORED_DEFAULT_WILDCARDS
+from ggshield.utils.git_shell import EMPTY_SHA, EMPTY_TREE
 from ggshield.verticals.secret.repo import cd
 from tests.repository import Repository
 from tests.unit.conftest import assert_invoke_exited_with, assert_invoke_ok
